Reorganization of the DQ utility classes
