+++ Analysis/CMakeLists.txt
@@ -12,3 +12,4 @@ add_subdirectory(Core)
 add_subdirectory(DataModel)
 add_subdirectory(Tasks)
 add_subdirectory(Tutorials)
+add_subdirectory(PWGDQ)
