+++ Analysis/PWGDQ/src/AnalysisCompositeCut.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "AnalysisCompositeCut.h"
+#include "PWGDQCore/AnalysisCompositeCut.h"
 
 ClassImp(AnalysisCompositeCut)
