+++ Analysis/PWGDQ/src/AnalysisCut.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "AnalysisCut.h"
+#include "PWGDQCore/AnalysisCut.h"
 
 ClassImp(AnalysisCut);
