+++ Analysis/PWGDQ/src/VarManager.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "VarManager.h"
+#include "PWGDQCore/VarManager.h"
 
 #include <TMath.h>
