+++ Analysis/Tasks/PWGDQ/CMakeLists.txt
@@ -1,18 +1,12 @@
-o2_add_library(PWGDQCore
-               SOURCES  VarManager.cxx
-                        HistogramManager.cxx
-                        AnalysisCut.cxx
-                        AnalysisCompositeCut.cxx
-               PUBLIC_LINK_LIBRARIES O2::Framework O2::AnalysisCore)
-
-o2_target_root_dictionary(PWGDQCore
-                          HEADERS  AnalysisCut.h 
-                          AnalysisCompositeCut.h 
-                          VarManager.h 
-                          HistogramManager.h
-                          CutsLibrary.h 
-                          HistogramsLibrary.h
-                          LINKDEF PWGDQCoreLinkDef.h)
+# Copyright CERN and copyright holders of ALICE O2. This software is distributed
+# under the terms of the GNU General Public License v3 (GPL Version 3), copied
+# verbatim in the file "COPYING".
+#
+# See http://alice-o2.web.cern.ch/license for full licensing information.
+#
+# In applying this license CERN does not waive the privileges and immunities
+# granted to it by virtue of its status as an Intergovernmental Organization or
+# submit itself to any jurisdiction.
 
 o2_add_dpl_workflow(table-maker
                     SOURCES tableMaker.cxx
