+++ Analysis/Tasks/PWGDQ/filterPP.cxx
@@ -19,12 +19,12 @@
 #include "AnalysisDataModel/Centrality.h"
 #include "AnalysisCore/TriggerAliases.h"
 #include "AnalysisDataModel/ReducedInfoTables.h"
-#include "VarManager.h"
-#include "HistogramManager.h"
-#include "AnalysisCut.h"
-#include "AnalysisCompositeCut.h"
-#include "CutsLibrary.h"
-#include "HistogramsLibrary.h"
+#include "PWGDQCore/VarManager.h"
+#include "PWGDQCore/HistogramManager.h"
+#include "PWGDQCore/AnalysisCut.h"
+#include "PWGDQCore/AnalysisCompositeCut.h"
+#include "PWGDQCore/CutsLibrary.h"
+#include "PWGDQCore/HistogramsLibrary.h"
 #include "AnalysisDataModel/PID/PIDResponse.h"
 #include "AnalysisDataModel/TrackSelectionTables.h"
 #include <TH1F.h>
