+++ Analysis/Tasks/PWGDQ/tableMaker_pp.cxx
@@ -19,10 +19,10 @@
 #include "AnalysisDataModel/Centrality.h"
 #include "AnalysisCore/TriggerAliases.h"
 #include "AnalysisDataModel/ReducedInfoTables.h"
-#include "VarManager.h"
-#include "HistogramManager.h"
-#include "AnalysisCut.h"
-#include "AnalysisCompositeCut.h"
+#include "PWGDQCore/VarManager.h"
+#include "PWGDQCore/HistogramManager.h"
+#include "PWGDQCore/AnalysisCut.h"
+#include "PWGDQCore/AnalysisCompositeCut.h"
 #include "AnalysisDataModel/PID/PIDResponse.h"
 #include "AnalysisDataModel/TrackSelectionTables.h"
 #include <iostream>
